requirements.txt
opencv-python==4.8.0.74
face-recognition==1.3.0
pillow==10.0.0
twilio==7.18.0
ttkthemes==3.2.2
cmake==3.26.4
dlib==19.23.0
ttk
